Move ChatRepository to :feature-chat
