--- app/app/src/main/kotlin/com/hedvig/app/util/extensions/InputStreamExtensions.kt
@@ -1,10 +1,0 @@
-package com.hedvig.app.util.extensions
-
-import java.io.File
-import java.io.InputStream
-
-fun InputStream.into(file: File) {
-  use { inputStream ->
-    file.outputStream().use { inputStream.copyTo(it) }
-  }
-}
