+++ app/feature/feature-chat/build.gradle.kts
@@ -11,15 +11,18 @@ dependencies {
   implementation(libs.androidx.navigation.compose)
   implementation(libs.androidx.navigation.runtime)
   implementation(libs.androidx.other.fragment)
+  implementation(libs.apollo.normalizedCache)
   implementation(libs.coroutines.core)
   implementation(libs.kiwi.navigationCompose)
   implementation(libs.koin.compose)
   implementation(libs.koin.core)
   implementation(libs.kotlinx.serialization.core)
+  implementation(projects.apolloCore)
   implementation(projects.apolloGiraffePublic)
   implementation(projects.coreBuildConstants)
   implementation(projects.coreCommonPublic)
   implementation(projects.coreDemoMode)
+  implementation(projects.hanalyticsFeatureFlagsPublic)
   implementation(projects.languageCore)
   implementation(projects.navigationActivity)
   implementation(projects.navigationComposeTyped)
